+++ packages/react-openapi/src/index.ts
@@ -1,5 +1,5 @@
 export * from './models';
 export * from './OpenAPIOperation';
 export * from './OpenAPIOperationContext';
-export { resolveOpenAPIOperation } from './resolveOpenAPIOperation';
+export * from './resolveOpenAPIOperation';
 export type { OpenAPIModelsData, OpenAPIOperationData } from './types';
