+++ packages/react-openapi/src/models/index.ts
@@ -1,2 +1,2 @@
-export { OpenAPIModels } from './OpenAPIModels';
+export * from './OpenAPIModels';
 export * from './resolveOpenAPIModels';
